完善 Leader 选举的代码

+++ core/src/main/java/zbl/moonlight/core/raft/log/RaftLog.java
@@ -123,7 +123,7 @@ public class RaftLog {
     private synchronized EntryIndex getEntryIndexByIndex(int index) throws IOException {
         assert index > 0;
         ByteBuffer buffer = ByteBuffer.allocate(ENTRY_INDEX_LENGTH);
-        indexFile.read(buffer, ((long) index) * ENTRY_INDEX_LENGTH);
+        indexFile.read(buffer, ((long) index) * (long) ENTRY_INDEX_LENGTH);
         return EntryIndex.fromBytes(buffer.array());
     }
